fix the VM sizes map
Some VM sizes are available on Azure and Azure Stack.
For them exclusively:
* the storageAccountType property value has been set to "Standard_LRS"
  and the storageAccountTier has been set to "Standard".
* the acceleratedNetworking property value has been set to "false".

Because in many cases it does not make sense to offer Premium_LRS on
Azure Stack because most of the time the hardware will have Standard
HDD (with SSD for caching features) and moving from Standard_LRS to
Premium_LRS do nothing except increase the billing.
Then Cloud Ops set the Quota to 0 Premium disks to prevent this from
happening.

BTW acceleratedNetworking is not available on Azure Stack that's why
this matching property value has been also fixed.