Add support for 2 LB in front of the Masters
In production it's recommended to use 2 Load Balancers in front of the
Masters.
The private Load Balancer will be used internally by the OpenShift nodes
to query API.
And the public Load Balancer will be used externally for the Webconsole.

By the way it's the required configuration to use Custom Certificates
(aka named certificates) for the Masters.
See http://v1.uncontained.io/playbooks/installation/load_balancing.html
for more details.

To manage Load Balancer limitations that prevent VM participating to the
backend pool to access the internal Load Balancer frontend IP, a
specific dnsmasq configuration will be deployed on each master to make
them querying themself instead.
See
https://docs.microsoft.com/en-us/azure/load-balancer/load-balancer-troubleshoot#cause-4-accessing-the-internal-load-balancer-frontend-from-the-participating-load-balancer-backend-pool-vm
for more details.

If for any reason the automatic installation of OpenShift is disabled by
setting "automaticDeployment" to "manual", this dnsmasq settings would
be mandatory and would have to be made handly.