try to add internal azure DNS zone
The main aim is to manage the name of the internal api LB.
Because if we want to override it on master, we can't use directly an IP
address.
Then we will create an azure dns zone that will contain all the node
names and we will add the masterAPIDNS name.
In a first time we commit the pipeline without the automatic deployment
to ensure the DNS zone is rightly created.